requirements.txt
click==8.1.7
minimalmodbus==2.1.1
pika==1.3.2
pyserial==3.5
smbus2==0.4.3
termcolor==2.4.0
w1thermsensor==2.3.0
